1.2.16 Rational

+++ src/main/scala/one/one/Euclid.scala
@@ -6,7 +6,7 @@ object Euclid {
     println(rec(1234567, 1111111))
   }
 
-  def doAlg(a: Int, b: Int): Int = {
+  def doAlg(a: Long, b: Long): Long = {
     if (a == 0 || b == 0) 1 else {
       if (b > a) rec(b, a)
       else rec(a, b)
@@ -19,7 +19,7 @@ object Euclid {
     * @param q
     * @return
     */
-  private def rec(p: Int, q: Int): Int = {
+  private def rec(p: Long, q: Long): Long = {
 //    println(s"$p, $q")
     val o = p % q
     if (o == 0) q
